// src/app/models/personalInfo.ts
//  This class is responsible for declare Schema User Personal Info

export class PersonalInfo {

  apartment?: string;
  city?: string;
  country?: string;
  dateOfBirth?: string | string;
  firstName?: string;
  languagePreference?: string;
  lastName?: string;
  personalInfoId?: number;
  postalCode?: string;
  primaryTelephoneNumber?: string;
  provinceOrState?: string;
  secondaryTelephoneNumber?: string;
  socialInsuranceNumber?: string;
  streetAddress?: string;
  otherLanguagePreference?: string;

  constructor() { }
}